Remove unneeded Mockaroo stuff, show selected platform in dropdown
