+++ front-end/src/DropdownOptions.js
@@ -15,7 +15,34 @@ const statuses = [
   { value: 'In Progress', label: 'In Progress' }
 ]
 
+// textToValue transforms a text value to a value/label format to pre-populate
+// the Select component that calls the function. It currently supports both
+// platform dropdowns and status dropdowns
+const textToValue = (text, type) => {
+  let array = []
+  switch (type) {
+    case 'platform':
+      array = platforms
+      break
+    case 'status':
+      array = statuses
+      break
+    default:
+      break
+  }
+  const match = array.filter((p) => p.value === text)
+  if (match.length === 0) {
+    return {
+      value: '',
+      label: `Select a ${type.charAt(0).toUpperCase() + type.slice(1)}`
+    }
+  } else {
+    return match[0]
+  }
+}
+
 module.exports = {
   platforms: platforms,
-  statuses: statuses
+  statuses: statuses,
+  textToValue: textToValue
 }
